sm8550-common: parts: Remove GestureUtils
Using a commit from AOSPA we were able to get hardware/xiaomi working correctly and writing gestures nodes as it should. Now when the device is off, nodes that the user has enabled are set to 1. The device is on, the nodes are turned off. This is important for NonUI Notifier to work.

https://github.com/AOSPA/android_hardware_xiaomi/commit/a0a0bb6485890e75abff8672c25fdbf8c98f1e4b